Silenced compiler warnings
Silenced most uneeded compiler warnings in pep9cpu.

Still in the process of silencing remaining errors, however most that were left stem from questionable behavior.

+++ pep9cpu/microcode.cpp
@@ -301,7 +301,7 @@ void MicroCode::setFalseTarget(const SymbolEntry* target)
     falseTargetAddr = target;
 }
 
-int MicroCode::getControlSignal(Enu::EControlSignals field) const
+quint8 MicroCode::getControlSignal(Enu::EControlSignals field) const
 {
     return controlSignals[field];
 }
